0.1: library version
