+++ README.md
@@ -2,16 +2,11 @@
 Preprinting preparation tool for PDF ebooks.  
 
 ### Build and Install
-Enter directory src  
-`cd src`  
-run  
-```
+Enter project directory and type:
+```sh
 make -j4  
 sudo make install  
 ```
-Install manpage  
-`sudo make installman`  
-
 
 ### Features
 * PDF v1.7 support  
@@ -27,13 +22,26 @@ Install manpage
 See manual page (PDF or man page) for detailed usage  
 
 Scale to print in A4 size paper  
-`pdfcook 'scaleto(a4)' input.pdf output.pdf`  
+```sh
+pdfcook 'scaleto(a4)' input.pdf output.pdf  
+```
 
 Add binding margin after scaling (? for odd pages, + for even pages)  
-`pdfcook 'scaleto(a4) move(20){?} move(-20){+}' input.pdf output.pdf`  
+```sh
+pdfcook 'scaleto(a4) move(20){?} move(-20){+}' input.pdf output.pdf  
+```
 
 Add page numbers  
-`pdfcook 'number' input.pdf output.pdf`  
+```sh
+pdfcook 'number' input.pdf output.pdf  
+```
 
 Booklet format  
-`pdfcook 'book nup(2, paper=a4)' input.pdf output.pdf`  
+```sh
+pdfcook 'book nup(2, paper=a4)' input.pdf output.pdf  
+```
+
+----
+
+2021
+
