+++ src/include/geometry.h
@@ -12,7 +12,8 @@ public:
     Point(float x, float y);
 };
 
-class Rect {
+class Rect
+{
 public:
     Point left; // bottom left (0,0)
     Point right;// top right
