+++ src/include/pdf_objects.h
@@ -3,6 +3,7 @@
 #include <vector>
 #include <map>
 #include <set>
+#include <string>
 #include "fileio.h"
 
 #define PDF_NAME_MAX_LEN 255
@@ -21,7 +22,8 @@ class Token;
 class PdfObject;
 class ObjectTable;
 
-typedef struct {
+typedef struct
+{
     char *data;
     int len;// string length excluding null character
 } String;
@@ -71,7 +73,8 @@ public:
     int         write (FILE *f);
     MapIter     begin();
     MapIter     end();
-    PdfObject* operator[] (std::string key) {
+    PdfObject* operator[] (std::string key)
+    {
         if (dict.count(key) < 1)
             return NULL;
         return dict[key];
@@ -95,7 +98,8 @@ public:
     ~StreamObj();
 };
 
-typedef struct {
+typedef struct
+{
     int major;
     int minor;
     PdfObject *obj;
@@ -103,7 +107,8 @@ typedef struct {
 
 
 
-typedef enum {
+typedef enum
+{
     PDF_OBJ_BOOL, PDF_OBJ_INT, PDF_OBJ_REAL, PDF_OBJ_STR,
     PDF_OBJ_NAME, PDF_OBJ_ARRAY, PDF_OBJ_DICT, PDF_OBJ_STREAM,
     PDF_OBJ_INDIRECT, PDF_OBJ_INDIRECT_REF, PDF_OBJ_NULL, PDF_OBJ_UNKNOWN
@@ -114,7 +119,8 @@ class PdfObject
 {
 public:
     ObjectType type;
-    union {
+    union
+    {
         BoolObj     boolean;
         IntObj      integer;
         RealObj     real;
@@ -142,24 +148,26 @@ public:
 
 
 
-typedef struct {
-	PdfObject *obj;
+typedef struct
+{
+    PdfObject *obj;
     int8_t type;    // 0=free(f), 1=nonfree(n), 2=compressed
-	int major;   // object no.
-	int minor; // gen id for type 1, (always 0 for type 2)
-    union {
+    int major;   // object no.
+    int minor; // gen id for type 1, (always 0 for type 2)
+    union
+    {
         int offset; // offset of obj from beginning of file (type 1 only)
         int next_free;// obj no of next free obj (type 0 only)
         int obj_stm; // obj no. of object stream where obj is stored (type 2 only)
     };
     int index;// index no. within the obj stream (for type 2)
-	bool used;
+    bool used;
 } ObjectTableItem;
 
 class ObjectTable
 {
 public:
-	std::vector<ObjectTableItem> table;
+    std::vector<ObjectTableItem> table;
 
     int count();
     void expandToFit(size_t size);
@@ -172,7 +180,8 @@ public:
     void writeObjects(FILE *f);
     void writeXref (FILE *f);
 
-    ObjectTableItem& operator[] (int index) {
+    ObjectTableItem& operator[] (int index)
+    {
         return table[index];
     }
 };
@@ -182,7 +191,8 @@ public:
 enum  { CHAR_NULL=0, CHAR_TAB=9, CHAR_LF=10, CHAR_FF=12, CHAR_CR=13, CHAR_SP=32};
 
 /* constants identifying the token type */
-typedef enum {
+typedef enum
+{
     TOK_INT, TOK_REAL, TOK_STR, TOK_NAME, TOK_BDICT, TOK_EDICT,
     TOK_BARRAY, TOK_EARRAY, TOK_ID, TOK_EOF, TOK_UNKNOWN
 } TokType;
