+++ src/lib/pdf_objects.cpp
@@ -8,25 +8,31 @@
 
 // *********** ------------- Array Object ----------------- ***********
 //allows range based for loop
-ArrayIter ArrayObj:: begin() {
+ArrayIter ArrayObj:: begin()
+{
     return array.begin();
 }
-ArrayIter ArrayObj:: end() {
+ArrayIter ArrayObj:: end()
+{
     return array.end();
 }
-int ArrayObj:: count () {
+int ArrayObj:: count ()
+{
     return array.size();
 }
-PdfObject* ArrayObj:: at (int index) {
+PdfObject* ArrayObj:: at (int index)
+{
     return array[index];
 }
-void ArrayObj:: append (PdfObject *item) {
+void ArrayObj:: append (PdfObject *item)
+{
     array.push_back(item);
 }
 
 void ArrayObj:: deleteItems()
 {
-    for (PdfObject *item : array){
+for (PdfObject *item : array)
+    {
         delete item;
     }
     array.clear();
@@ -35,7 +41,8 @@ void ArrayObj:: deleteItems()
 int ArrayObj:: write (FILE *f)
 {
     int ret_val = fprintf(f, "[ ");
-    for (PdfObject *obj : this->array){
+for (PdfObject *obj : this->array)
+    {
         ret_val = obj->write(f);
         ret_val = fprintf(f, " ");
     }
@@ -44,30 +51,36 @@ int ArrayObj:: write (FILE *f)
 }
 
 // *********** ------------ Dictionary Object -------------- ***********
-void DictObj:: setDict (std::map<std::string, PdfObject*> &map){
+void DictObj:: setDict (std::map<std::string, PdfObject*> &map)
+{
     this->dict = map;
 }
 
-bool DictObj:: contains (std::string key) {
+bool DictObj:: contains (std::string key)
+{
     return (dict.count(key) > 0);
 }
 
-PdfObject* DictObj:: get (std::string key) {
+PdfObject* DictObj:: get (std::string key)
+{
     if (dict.count(key) < 1)
         return NULL;
     return dict[key];
 }
 
-void DictObj:: add (std::string key, PdfObject *val) {
+void DictObj:: add (std::string key, PdfObject *val)
+{
     dict[key] = val;
 };
 
 PdfObject* DictObj:: newItem (std::string key)
 {
-    if (dict.count(key) > 0) {
+    if (dict.count(key) > 0)
+    {
         dict[key]->clear();
     }
-    else {
+    else
+    {
         dict[key] = new PdfObject();
     }
     return dict[key];
@@ -76,10 +89,12 @@ PdfObject* DictObj:: newItem (std::string key)
 // hard copy all items from src_dict to this, overwrite if exists
 void DictObj:: merge(DictObj *src_dict)
 {
-    for (auto it : src_dict->dict) {
+for (auto it : src_dict->dict)
+    {
         // if val of key is dict obj, the merge the dicts
         if (this->contains(it.first) && dict[it.first]->type==PDF_OBJ_DICT &&
-                                        it.second->type==PDF_OBJ_DICT) {
+                it.second->type==PDF_OBJ_DICT)
+        {
             dict[it.first]->dict->merge(it.second->dict);
             continue;
         }
@@ -90,11 +105,13 @@ void DictObj:: merge(DictObj *src_dict)
 
 void DictObj:: filter(DictFilter &filter_set)
 {
-    for (auto it=dict.begin(); it!=dict.end();) {
+    for (auto it=dict.begin(); it!=dict.end();)
+    {
         std::string key = it->first;
         PdfObject *val = it->second;
         it++;//it must be placed before dict.erase()
-        if (filter_set.count(key) == 0) {
+        if (filter_set.count(key) == 0)
+        {
             delete val;
             dict.erase(key);
         }
@@ -103,7 +120,8 @@ void DictObj:: filter(DictFilter &filter_set)
 
 void DictObj:: deleteItem (std::string key)
 {
-    if (dict.count(key) > 0) {
+    if (dict.count(key) > 0)
+    {
         PdfObject *val = dict[key];
         delete val;
         dict.erase(key);
@@ -112,7 +130,8 @@ void DictObj:: deleteItem (std::string key)
 
 void DictObj:: deleteItems()
 {
-    for (auto it : dict) {
+for (auto it : dict)
+    {
         delete it.second;
     }
     dict.clear();
@@ -122,7 +141,8 @@ int DictObj:: write (FILE *f)
 {
     fprintf(f, "<<\n");
 
-    for (auto it : dict){
+for (auto it : dict)
+    {
         fprintf(f, "/%s ", it.first.c_str());
         PdfObject *val = it.second;
         val->write(f);
@@ -132,16 +152,19 @@ int DictObj:: write (FILE *f)
     return 0;
 }
 
-MapIter DictObj:: begin() {
+MapIter DictObj:: begin()
+{
     return dict.begin();
 }
-MapIter DictObj:: end() {
+MapIter DictObj:: end()
+{
     return dict.end();
 }
 
 // *********** ------------- Stream Object ----------------- ***********
 
-StreamObj:: StreamObj() {
+StreamObj:: StreamObj()
+{
     stream = NULL;
     len = 0;
 }
@@ -156,8 +179,10 @@ int StreamObj:: write (FILE *f)
     fprintf(f,"\nstream\n");
 
     assert (this->stream!=NULL);
-    if (this->len){
-        if (fwrite(this->stream, sizeof(char), this->len, f) != this->len){
+    if (this->len)
+    {
+        if (fwrite(this->stream, sizeof(char), this->len, f) != this->len)
+        {
             message(FATAL, "fwrite() error");
         }
     }
@@ -168,25 +193,30 @@ int StreamObj:: write (FILE *f)
 bool StreamObj:: decompress()
 {
     PdfObject *p_obj = this->dict["Filter"];
-    if (p_obj == NULL) {
+    if (p_obj == NULL)
+    {
         return true;
     }
-    switch (p_obj->type){
+    switch (p_obj->type)
+    {
     case PDF_OBJ_ARRAY:
+    {
+for (PdfObject *filter : *p_obj->array)
         {
-        for (PdfObject *filter : *p_obj->array){
             assert(filter->type==PDF_OBJ_NAME);
             /*while (filter->type == PDF_OBJ_INDIRECT_REF){ // FIXME if error occurs
                 // get the object from reference
             }*/
-            if (apply_decompress_filter(filter->name, &(this->stream), &(this->len), this->dict) != 0){
+            if (apply_decompress_filter(filter->name, &(this->stream), &(this->len), this->dict) != 0)
+            {
                 return false;
             }
         }
         break;
     }
     case PDF_OBJ_NAME:
-        if (apply_decompress_filter(p_obj->name, &(this->stream), &(this->len), this->dict) != 0){
+        if (apply_decompress_filter(p_obj->name, &(this->stream), &(this->len), this->dict) != 0)
+        {
             message(FATAL, "failed to decompress object");
         }
         break;
@@ -200,45 +230,51 @@ bool StreamObj:: decompress()
 
 bool StreamObj:: compress (const char *filter)
 {
-	char *ch;
+    char *ch;
 
-	if (apply_compress_filter(filter, &(this->stream), &(this->len), this->dict) != 0){
-		return false;
-	}
-	PdfObject *filter_obj = this->dict.get("Filter");
+    if (apply_compress_filter(filter, &(this->stream), &(this->len), this->dict) != 0)
+    {
+        return false;
+    }
+    PdfObject *filter_obj = this->dict.get("Filter");
 
-	if (!filter_obj) {
-		filter_obj = this->dict.newItem("Filter");
-		asprintf(&ch,"/%s",filter);
-		filter_obj->readFromString(ch);
-		free(ch);
-	}
-	else {// already contains a filter
-	  	switch (filter_obj->type){
-		case PDF_OBJ_ARRAY:
-            {
-			asprintf(&ch,"/%s",filter);
+    if (!filter_obj)
+    {
+        filter_obj = this->dict.newItem("Filter");
+        asprintf(&ch,"/%s",filter);
+        filter_obj->readFromString(ch);
+        free(ch);
+    }
+    else  // already contains a filter
+    {
+        switch (filter_obj->type)
+        {
+        case PDF_OBJ_ARRAY:
+        {
+            asprintf(&ch,"/%s",filter);
             PdfObject *array_item = new PdfObject();
-			array_item->readFromString(ch);
-			free(ch);
-			filter_obj->array->append(array_item);
-            }
-			break;
-		case PDF_OBJ_NAME:
-			asprintf(&ch, " [ /%s /%s ] ", filter_obj->name, filter);
-			filter_obj->clear();
-			filter_obj->readFromString(ch);
-			free(ch);
-			break;
-		default:
-			assert(0);
-		}
-	}
-	return true;
+            array_item->readFromString(ch);
+            free(ch);
+            filter_obj->array->append(array_item);
+        }
+        break;
+        case PDF_OBJ_NAME:
+            asprintf(&ch, " [ /%s /%s ] ", filter_obj->name, filter);
+            filter_obj->clear();
+            filter_obj->readFromString(ch);
+            free(ch);
+            break;
+        default:
+            assert(0);
+        }
+    }
+    return true;
 }
 
-StreamObj:: ~StreamObj() {
-    if (stream!=NULL){
+StreamObj:: ~StreamObj()
+{
+    if (stream!=NULL)
+    {
         free(stream);
     }
     dict.deleteItems();
@@ -247,7 +283,8 @@ StreamObj:: ~StreamObj() {
 
 // *********** -------------- Pdf Object ----------------- ***********
 
-PdfObject:: PdfObject() {
+PdfObject:: PdfObject()
+{
     type = PDF_OBJ_UNKNOWN;
 }
 
@@ -276,7 +313,8 @@ bool
 PdfObject:: readFromString (const char *str)
 {
     MYFILE *f = stropen(str);
-    if (f==NULL){
+    if (f==NULL)
+    {
         return false;
     }
     bool retval = this->read(f, NULL, NULL);
@@ -288,29 +326,35 @@ PdfObject:: readFromString (const char *str)
 bool
 PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
 {
-    uint stream_len = 0;
+    size_t stream_len = 0;
     PdfObject *new_obj;
     std::map<std::string, PdfObject*>  new_dict;
     size_t fpos;
     Token tok;
-    if (last_tok==NULL){
+    if (last_tok==NULL)
+    {
         last_tok=&tok;
     }
     //printf("get obj %ld\n", myftell(f));
     //this->type=PDF_OBJ_UNKNOWN;//see constructor
-    while (last_tok->get(f)){
-        switch (last_tok->type){
+    while (last_tok->get(f))
+    {
+        switch (last_tok->type)
+        {
         case TOK_INT://maybe integer, indirect, or indirect reference obj
             this->setType(PDF_OBJ_INT);
             this->integer = last_tok->integer;
-            if (last_tok->sign){//it is integer, not indirect object
+            if (last_tok->sign) //it is integer, not indirect object
+            {
                 return true;
             }
             fpos = myftell(f);
             last_tok->get(f);
-            if (last_tok->type!=TOK_INT || last_tok->sign){// not indirect object
+            if (last_tok->type!=TOK_INT || last_tok->sign) // not indirect object
+            {
                 last_tok->freeData();
-                if ( myfseek(f, fpos, SEEK_SET)==EOF ){
+                if ( myfseek(f, fpos, SEEK_SET)==EOF )
+                {
                     message(FATAL,"myfseek()  error in file %s at line %d", __FILE__, __LINE__);
                     return false;
                 }
@@ -318,9 +362,11 @@ PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
             }
             // we have two integers, check if there is 'obj' or 'R' next to it
             last_tok->get(f);
-            if (last_tok->type!=TOK_ID){
+            if (last_tok->type!=TOK_ID)
+            {
                 last_tok->freeData();
-                if (myfseek(f,fpos,SEEK_SET)==EOF){
+                if (myfseek(f,fpos,SEEK_SET)==EOF)
+                {
                     message(FATAL,"myfseek()  error at in file %s line  %d",__FILE__, __LINE__);
                     return false;
                 }
@@ -329,25 +375,30 @@ PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
             this->indirect.major = this->integer;
             this->indirect.minor = last_tok->integer;
 
-            if (strcmp(last_tok->id,"R")==0){
+            if (strcmp(last_tok->id,"R")==0)
+            {
                 this->setType(PDF_OBJ_INDIRECT_REF);
                 return true;
             }
-            if (strcmp(last_tok->id,"obj")==0){
+            if (strcmp(last_tok->id,"obj")==0)
+            {
                 this->setType(PDF_OBJ_INDIRECT);
                 this->indirect.obj = new PdfObject();
-                if (not this->indirect.obj->read(f,xref,last_tok)){
+                if (not this->indirect.obj->read(f,xref,last_tok))
+                {
                     message(WARN, "read indirect obj failed : ID %d %d", indirect.major, indirect.minor);
                     return false;
                 }
                 last_tok->get(f);
-                if (last_tok->type!=TOK_ID || strcmp(last_tok->id,"endobj")!=0){
+                if (last_tok->type!=TOK_ID || strcmp(last_tok->id,"endobj")!=0)
+                {
                     message(WARN, "endobj keyword not found : indirect obj %d %d", indirect.major, indirect.minor);
                 }
                 return true;
             }
             // two int numbers and a TOK_ID next to it other than 'R' and 'obj'
-            if (myfseek(f,fpos,SEEK_SET)==EOF){
+            if (myfseek(f,fpos,SEEK_SET)==EOF)
+            {
                 message(FATAL,"myfseek()  error in file %s at line %d",__FILE__,__LINE__);
                 return false;
             }
@@ -366,18 +417,21 @@ PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
             //this->str.type = last_tok->str.type;
             return true;
         case TOK_BDICT:// dictionary or stream obj
-            while (1) {
+            while (1)
+            {
                 if (!last_tok->get(f) or last_tok->type!=TOK_NAME)// get key
                     break;
                 std::string key(last_tok->name);
                 // get value of key
                 new_obj = new PdfObject();
-                if (not new_obj->read(f, xref, last_tok)){
+                if (not new_obj->read(f, xref, last_tok))
+                {
                     // now we can't further read, this should fail all its parent objects
                     delete new_obj;
                     if (this->type==PDF_OBJ_STREAM)// this will help deleting new_dict items
                         this->stream->dict.setDict(new_dict);
-                    else {
+                    else
+                    {
                         this->setType(PDF_OBJ_DICT);
                         this->dict->setDict(new_dict);
                     }
@@ -386,7 +440,8 @@ PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
                 }
                 // if dict has /Length key then it is stream object
                 // if stream length is indirect obj, get length as integer
-                if (key == "Length"){
+                if (key == "Length")
+                {
                     this->setType(PDF_OBJ_STREAM);
                     switch (new_obj->type)
                     {
@@ -394,19 +449,21 @@ PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
                         stream_len = new_obj->integer;
                         break;
                     case PDF_OBJ_INDIRECT_REF:
+                    {
+                        fpos = myftell(f);
+                        xref->readObject(f, new_obj->indirect.major);
+                        PdfObject *ref_obj = xref->table[new_obj->indirect.major].obj;
+                        if (ref_obj->type == PDF_OBJ_INT)
                         {
-                            fpos = myftell(f);
-                            xref->readObject(f, new_obj->indirect.major);
-                            PdfObject *ref_obj = xref->table[new_obj->indirect.major].obj;
-                            if (ref_obj->type == PDF_OBJ_INT) {
-                                stream_len = ref_obj->integer;
-                            }
-                            else {
-                                message(FATAL, "Stream length is not int");
-                            }
-                            myfseek(f, fpos, SEEK_SET);
+                            stream_len = ref_obj->integer;
                         }
-                        break;
+                        else
+                        {
+                            message(FATAL, "Stream length is not int");
+                        }
+                        myfseek(f, fpos, SEEK_SET);
+                    }
+                    break;
                     default:
                         message(WARN, "Can't read stream length of type %d", new_obj->type);
                         return false;
@@ -416,55 +473,66 @@ PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
                 else // dont add Length key for stream, it will be added while writing to file
                     new_dict[key] = new_obj;
             }
-            if (last_tok->type!=TOK_EDICT){
+            if (last_tok->type!=TOK_EDICT)
+            {
                 message(WARN, "dictionary does not have end bracket");
                 return false;
             }
-            if (this->type!=PDF_OBJ_STREAM) {
+            if (this->type!=PDF_OBJ_STREAM)
+            {
                 this->setType(PDF_OBJ_DICT);
                 this->dict->setDict(new_dict);
                 return true;
             }
             this->stream->dict.setDict(new_dict);
             if ( (not last_tok->get(f))
-                || last_tok->type!=TOK_ID
-                || strcmp(last_tok->id, "stream")!=0) {
+                    || last_tok->type!=TOK_ID
+                    || strcmp(last_tok->id, "stream")!=0)
+            {
                 message(WARN, "stream keyword not found");
                 return false;
             }
-            switch (mygetc(f)){
-                case EOF:
-                    return false;
-                case CHAR_CR:
-                    if (mygetc(f)!=CHAR_LF){
-                        myungetc(f);
-                    }
-                case CHAR_LF:
-                    break;
-                default:
+            switch (mygetc(f))
+            {
+            case EOF:
+                return false;
+            case CHAR_CR:
+                if (mygetc(f)!=CHAR_LF)
+                {
                     myungetc(f);
-                    break;
+                }
+            case CHAR_LF:
+                break;
+            default:
+                myungetc(f);
+                break;
             }
             this->stream->begin = myftell(f);
             this->stream->len = stream_len;
-            if (stream_len){
+            if (stream_len)
+            {
                 this->stream->stream = (char *)malloc(sizeof(char) * stream_len);
-                if (this->stream->stream==NULL){
+                if (this->stream->stream==NULL)
+                {
                     message(FATAL,"malloc() error");
                 }
-                if (myfread(this->stream->stream,sizeof(char),stream_len,f)!=stream_len){
+                if (myfread(this->stream->stream,sizeof(char),stream_len,f)!=stream_len)
+                {
                     message(FATAL,"fread() error");
                 }
             }
-            else {// for stream length is 0
+            else  // for stream length is 0
+            {
                 this->stream->stream = (char *)malloc(sizeof(char) * 1);
-                if (this->stream->stream==NULL){
+                if (this->stream->stream==NULL)
+                {
                     message(FATAL,"malloc() error");
                 }
             }
             if (not last_tok->get(f)
-                || last_tok->type!=TOK_ID
-                || strcmp(last_tok->id,"endstream")!=0){
+                    || last_tok->type!=TOK_ID
+                    || strcmp(last_tok->id,"endstream")!=0)
+            {
                 message(WARN, "endstream keyword not found");
                 return false;
             }
@@ -472,27 +540,32 @@ PdfObject:: read (MYFILE *f, ObjectTable *xref, Token *last_tok)
         case TOK_BARRAY:
             this->setType(PDF_OBJ_ARRAY);
             new_obj = new PdfObject();
-            while (new_obj->read(f,xref,last_tok)){
+            while (new_obj->read(f,xref,last_tok))
+            {
                 this->array->append(new_obj);
                 new_obj = new PdfObject();
             }
             delete new_obj;
-            if (last_tok->type!=TOK_EARRAY){
+            if (last_tok->type!=TOK_EARRAY)
+            {
                 message(WARN, "Array : ending bracket not found");
                 return false;
             }
             return true;
         case TOK_ID:
-            if (strcmp(last_tok->id,"null")==0){
+            if (strcmp(last_tok->id,"null")==0)
+            {
                 this->setType(PDF_OBJ_NULL);
                 return true;
             }
-            if (strcmp(last_tok->id,"true")==0){
+            if (strcmp(last_tok->id,"true")==0)
+            {
                 this->setType(PDF_OBJ_BOOL);
                 this->boolean = true;
                 return true;
             }
-            if (strcmp(last_tok->id,"false")==0){
+            if (strcmp(last_tok->id,"false")==0)
+            {
                 this->setType(PDF_OBJ_BOOL);
                 this->boolean = false;
                 return true;
@@ -518,10 +591,12 @@ PdfObject:: write (FILE * f)
     switch (this->type)
     {
     case PDF_OBJ_BOOL:
-        if (this->boolean){
+        if (this->boolean)
+        {
             ret_val = fprintf(f, "true");
         }
-        else {
+        else
+        {
             ret_val = fprintf(f, "false");
         }
         return ret_val<0?ret_val:0;
@@ -562,81 +637,90 @@ PdfObject:: write (FILE * f)
 }
 
 int
-PdfObject:: copyFrom (PdfObject *src_obj){
+PdfObject:: copyFrom (PdfObject *src_obj)
+{
     // create deep copy of all objects
     this->setType(src_obj->type);
-    switch (src_obj->type){
-        case PDF_OBJ_BOOL:
-            this->boolean = src_obj->boolean;
-            return true;
-        case PDF_OBJ_INT:
-            this->integer = src_obj->integer;
-            return true;
-        case PDF_OBJ_REAL:
-            this->real = src_obj->real;
-            return true;
-        case PDF_OBJ_STR:
-            str.len = src_obj->str.len;
-            str.data = (char*) malloc( str.len+1);
-            assert(str.data!=NULL);
-            memcpy(str.data, src_obj->str.data, str.len+1);
-            return true;
-        case PDF_OBJ_NAME:
-            this->name = strdup(src_obj->name);
-            if (this->name==NULL)
+    switch (src_obj->type)
+    {
+    case PDF_OBJ_BOOL:
+        this->boolean = src_obj->boolean;
+        return true;
+    case PDF_OBJ_INT:
+        this->integer = src_obj->integer;
+        return true;
+    case PDF_OBJ_REAL:
+        this->real = src_obj->real;
+        return true;
+    case PDF_OBJ_STR:
+        str.len = src_obj->str.len;
+        str.data = (char*) malloc( str.len+1);
+        assert(str.data!=NULL);
+        memcpy(str.data, src_obj->str.data, str.len+1);
+        return true;
+    case PDF_OBJ_NAME:
+        this->name = strdup(src_obj->name);
+        if (this->name==NULL)
+            return false;
+        return true;
+    case PDF_OBJ_ARRAY:
+for (PdfObject *item : *src_obj->array)
+        {
+            PdfObject *new_item = new PdfObject();
+            new_item->copyFrom(item);
+            this->array->append(new_item);
+        }
+        return true;
+    case PDF_OBJ_DICT:
+for (auto it : *src_obj->dict)
+        {
+            PdfObject *new_obj = new PdfObject();
+            new_obj->copyFrom(it.second);
+            this->dict->add(it.first, new_obj);
+        }
+        return true;
+    case PDF_OBJ_STREAM:
+        this->stream->len = src_obj->stream->len;
+        // copy stream dictionary recursively
+for (auto it : src_obj->stream->dict)
+        {
+            PdfObject *new_obj = new PdfObject();
+            new_obj->copyFrom(it.second);
+            this->stream->dict.add(it.first, new_obj);
+        }
+        assert (src_obj->stream->stream!=NULL);
+        if (src_obj->stream->len)
+        {
+            this->stream->stream = (char *) malloc(sizeof(char) * (src_obj->stream->len));
+            if (this->stream->stream==NULL)
+            {
                 return false;
-            return true;
-        case PDF_OBJ_ARRAY:
-            for (PdfObject *item : *src_obj->array){
-                PdfObject *new_item = new PdfObject();
-                new_item->copyFrom(item);
-                this->array->append(new_item);
             }
-            return true;
-        case PDF_OBJ_DICT:
-            for (auto it : *src_obj->dict){
-                PdfObject *new_obj = new PdfObject();
-                new_obj->copyFrom(it.second);
-                this->dict->add(it.first, new_obj);
+            memcpy(this->stream->stream, src_obj->stream->stream, src_obj->stream->len);
+        }
+        else
+        {
+            this->stream->stream = (char *) malloc(sizeof(char) * 1);
+            if (this->stream->stream==NULL)
+            {
+                return false;
             }
-            return true;
-        case PDF_OBJ_STREAM:
-            this->stream->len = src_obj->stream->len;
-            // copy stream dictionary recursively
-            for (auto it : src_obj->stream->dict){
-                PdfObject *new_obj = new PdfObject();
-                new_obj->copyFrom(it.second);
-                this->stream->dict.add(it.first, new_obj);
-            }
-            assert (src_obj->stream->stream!=NULL);
-            if (src_obj->stream->len){
-                this->stream->stream = (char *) malloc(sizeof(char) * (src_obj->stream->len));
-                if (this->stream->stream==NULL){
-                    return false;
-                }
-                memcpy(this->stream->stream, src_obj->stream->stream, src_obj->stream->len);
-            }
-            else {
-                this->stream->stream = (char *) malloc(sizeof(char) * 1);
-                if (this->stream->stream==NULL){
-                    return false;
-                }
-            }
-            return true;
-        case PDF_OBJ_INDIRECT:
-            this->indirect.major = src_obj->indirect.major;
-            this->indirect.minor = src_obj->indirect.minor;
-            this->indirect.obj = new PdfObject();
-            this->indirect.obj->copyFrom(src_obj->indirect.obj);
-            return true;
-         case PDF_OBJ_INDIRECT_REF:
-            this->indirect.major = src_obj->indirect.major;
-            this->indirect.minor = src_obj->indirect.minor;
-            return true;
-        case PDF_OBJ_NULL:
-            return true;
-        default:
-            assert(0);
+        }
+        return true;
+    case PDF_OBJ_INDIRECT:
+        this->indirect.major = src_obj->indirect.major;
+        this->indirect.minor = src_obj->indirect.minor;
+        this->indirect.obj = new PdfObject();
+        this->indirect.obj->copyFrom(src_obj->indirect.obj);
+        return true;
+    case PDF_OBJ_INDIRECT_REF:
+        this->indirect.major = src_obj->indirect.major;
+        this->indirect.minor = src_obj->indirect.minor;
+        return true;
+    case PDF_OBJ_NULL:
+        return true;
+    default:
+        assert(0);
     }
     return true;
 }
@@ -671,20 +755,24 @@ void PdfObject:: clear()
     this->type = PDF_OBJ_UNKNOWN;
 }
 
-PdfObject:: ~PdfObject() {
+PdfObject:: ~PdfObject()
+{
     clear();
 }
 
 
 // *********** -------------- Pdf ObjectTable ----------------- ***********
 int
-ObjectTable:: count() {
+ObjectTable:: count()
+{
     return table.size();
 }
 
 void
-ObjectTable:: expandToFit (size_t size) {
-    if (size > table.size()) {
+ObjectTable:: expandToFit (size_t size)
+{
+    if (size > table.size())
+    {
         ObjectTableItem item = {NULL,0,0,0,0,0,0};
         table.resize(size, item);
     }
@@ -696,44 +784,53 @@ ObjectTable:: readObject(MYFILE *f, int major)
 {
     if (table[major].obj != NULL) return true;// already read
     // read object if nonfree object
-    if (table[major].type==1) {
+    if (table[major].type==1)
+    {
         PdfObject obj;
         int offset = table[major].offset;
-        if (myfseek(f, offset, SEEK_SET)){
+        if (myfseek(f, offset, SEEK_SET))
+        {
             message(WARN,"myfseek()  error, pos %d", offset);
             return false;
         }
-        if (not obj.read(f, this, NULL)){
+        if (not obj.read(f, this, NULL))
+        {
             message(WARN,"PdfObject::read() failed: nonfree obj %d", major);
             table[major].obj = new PdfObject();
             table[major].obj->type = PDF_OBJ_NULL;
             return false;
         }
-        if (obj.type!=PDF_OBJ_INDIRECT){
+        if (obj.type!=PDF_OBJ_INDIRECT)
+        {
             message(WARN, "Object %d isn't indirect", major);
             table[major].obj = new PdfObject();
             table[major].obj->type = PDF_OBJ_NULL;
             return false;
         }
-        if (obj.indirect.major!=major || obj.indirect.minor!=table[major].minor){
+        if (obj.indirect.major!=major || obj.indirect.minor!=table[major].minor)
+        {
             message(WARN, "Major or minor number in object are mismatched");
         }
         table[major].obj = obj.indirect.obj;
         obj.type = PDF_OBJ_UNKNOWN;// this is to prevent obj.indirect.obj from being deleted
         // decompress if it is compressed object stream
-        if (table[major].obj->type==PDF_OBJ_STREAM) {//todo : move to type==2 portion
+        if (table[major].obj->type==PDF_OBJ_STREAM)  //todo : move to type==2 portion
+        {
             PdfObject *type = table[major].obj->stream->dict["Type"];
-            if (type!=NULL && type->type==PDF_OBJ_NAME && strcmp(type->name,"ObjStm")==0){
+            if (type!=NULL && type->type==PDF_OBJ_NAME && strcmp(type->name,"ObjStm")==0)
+            {
                 table[major].obj->stream->decompress();
             }
         }
     }
     // read object if compressed nonfree object
-    else if (table[major].type==2) {
+    else if (table[major].type==2)
+    {
         // this object is inside a object stream.
         int obj_stm_no = table[major].obj_stm;
         this->readObject(f, obj_stm_no);
-        if (table[obj_stm_no].obj->type != PDF_OBJ_STREAM) {
+        if (table[obj_stm_no].obj->type != PDF_OBJ_STREAM)
+        {
             message(WARN, "source obj stream of obj %d is invalid", major);
             table[major].obj = new PdfObject();
             table[major].obj->type = PDF_OBJ_NULL;
@@ -746,7 +843,8 @@ ObjectTable:: readObject(MYFILE *f, int major)
         // stream contains : obj_no1 offset1 obj_no2 offset2 ... obj_1 obj2 ...
         MYFILE *file = streamopen(obj_stm->stream, obj_stm->len);
         Token tok;
-        for (int i=0; i<n; i++) {
+        for (int i=0; i<n; i++)
+        {
             tok.get(file);
             int obj_no = tok.integer;
             tok.get(file);
@@ -756,7 +854,8 @@ ObjectTable:: readObject(MYFILE *f, int major)
             size_t last_seek = myftell(file);
             myfseek(file, offset, SEEK_SET);
             PdfObject *new_obj = new PdfObject();
-            if (not new_obj->read(file, this, NULL)){
+            if (not new_obj->read(file, this, NULL))
+            {
                 message(WARN,"PdfObject::read() failed : compressed obj %d", obj_no);
                 new_obj->type = PDF_OBJ_NULL;
             }
@@ -776,28 +875,31 @@ ObjectTable:: readObject(MYFILE *f, int major)
 void ObjectTable:: readObjects(MYFILE *f)
 {
     // at first load nonfree objects and then decompress object streams
-	for (size_t i=1; i<table.size(); ++i) {
+    for (size_t i=1; i<table.size(); ++i)
+    {
         //message(LOG, "reading obj %d, type %d", i, xref->table[i].type);
-		switch (table[i].type) {
-            case 0:     // free obj
+        switch (table[i].type)
+        {
+        case 0:     // free obj
+            break;
+        case 1:     //nonfree obj
+            /* some bad xref table may have offset==0, or offset > file size*/
+            //message(LOG, "obj no. %d, offset %d", i, xref->table[i].offset);
+            if (table[i].offset==0)
+            {
+                debug("warning : offset of nonfree obj no %d is 0", i);
+                table[i].type = 0;
                 break;
-            case 1:     //nonfree obj
-                /* some bad xref table may have offset==0, or offset > file size*/
-                //message(LOG, "obj no. %d, offset %d", i, xref->table[i].offset);
-                if (table[i].offset==0){
-                    debug("warning : offset of nonfree obj no %d is 0", i);
-                    table[i].type = 0;
-                    break;
-                }
-                readObject(f, i);
-                break;
-            case 2:      // compressed obj
-                readObject(f, i);
-                table[i].type = 1;
-            default:
-                break;
-		}
-	}
+            }
+            readObject(f, i);
+            break;
+        case 2:      // compressed obj
+            readObject(f, i);
+            table[i].type = 1;
+        default:
+            break;
+        }
+    }
 }
 
 // from PDF 1.5 the xreftable can be a stream in an indirect object.
@@ -809,7 +911,8 @@ bool ObjectTable:: getFromStream (MYFILE *f, PdfObject *p_trailer)
     //fd = fopen("trailer", "wb");
     //fd = fopen("xref", "wb");
     PdfObject content;
-    if (not content.read(f, NULL, NULL)) {
+    if (not content.read(f, NULL, NULL))
+    {
         message(FATAL, "Unable to get xref from stream");
     }
     PdfObject *obj = content.indirect.obj;
@@ -824,13 +927,15 @@ bool ObjectTable:: getFromStream (MYFILE *f, PdfObject *p_trailer)
     // split stream into table, W parameter is array of length 3
     ArrayObj *w_arr_obj = p_trailer->dict->get("W")->array;
     int w_arr[3];
-    for (int i=0; i<3; ++i) {
+    for (int i=0; i<3; ++i)
+    {
         w_arr[i] = w_arr_obj->at(i)->integer;
     }
     int row_len = w_arr[0] + w_arr[1] + w_arr[2];
     // Index is array of pairs of integers. Each pair has obj number and obj count
     PdfObject *index = p_trailer->dict->get("Index");
-    if (index==NULL) {
+    if (index==NULL)
+    {
         index = new PdfObject();
         char s[24];
         snprintf(s, 23, "[ 0 %d ]", table_size);
@@ -838,12 +943,15 @@ bool ObjectTable:: getFromStream (MYFILE *f, PdfObject *p_trailer)
         p_trailer->dict->add("Index", index);
     }
     auto item = index->array->begin();
-    for (int i=0; item != index->array->end(); item++) {
+    for (int i=0; item != index->array->end(); item++)
+    {
         int first = (*item)->integer;
         item++;
         int count = (*item)->integer;
-        for (int major=first; major<first+count; major++) {
-            if (this->table[major].type!=0){//skip when already set by next xref table
+        for (int major=first; major<first+count; major++)
+        {
+            if (this->table[major].type!=0) //skip when already set by next xref table
+            {
                 i+=row_len;
                 continue;
             }
@@ -855,7 +963,8 @@ bool ObjectTable:: getFromStream (MYFILE *f, PdfObject *p_trailer)
             ObjectTableItem *elm = &(this->table[major]);
             elm->major = major;
             elm->type = field1;
-            switch (field1) {
+            switch (field1)
+            {
             case 0:// free objects
                 elm->next_free = field2;
                 elm->minor = field3;
@@ -884,41 +993,52 @@ bool ObjectTable:: get (MYFILE *f, size_t xref_poz, char *line, PdfObject *p_tra
     int len=0, object_id=0, object_count=0;
 
     ObjectTableItem *elm;
-    if (myfseek(f, xref_poz, SEEK_SET)==-1){
+    if (myfseek(f, xref_poz, SEEK_SET)==-1)
+    {
         return false;
     }
     // bad pdf may contain a newline before 'xref'
     char c;
-    do { c = mygetc(f); }
+    do
+    {
+        c = mygetc(f);
+    }
     while (isspace(c));
     myungetc(f);
-    if ((myfgets(line,LLEN,f,NULL))==EOF){
+    if ((myfgets(line,LLEN,f,NULL))==EOF)
+    {
         return false;
     }
-    if (!starts(line, "xref")) {
+    if (!starts(line, "xref"))
+    {
         myfseek(f, xref_poz, SEEK_SET);
         return this->getFromStream(f, p_trailer);
     }
     //FILE *fd = fopen("xref", "wb");
-    while ((pos = myftell(f)) && myfgets(line,LLEN,f,NULL)!=EOF && len>=0){
+    while ((pos = myftell(f)) && myfgets(line,LLEN,f,NULL)!=EOF && len>=0)
+    {
         char *entry = line;
         while (isspace(*entry)) // fixes for leading spaces in xref table
             entry++;
         len = strlen(entry)-1;
         if (len==-1) continue; // skip empty lines
-        while (len >= 0 && isspace((unsigned char)(entry[len]))){
+        while (len >= 0 && isspace((unsigned char)(entry[len])))
+        {
             entry[len] = 0;
             --len;
         }
-        if (strlen(entry)==XREF_ENT_LEN){
+        if (strlen(entry)==XREF_ENT_LEN)
+        {
             int field1, field2;
             char obj_type;
-            if (sscanf(entry,"%d %d %c", &field1, &field2, &obj_type)!=3){
+            if (sscanf(entry,"%d %d %c", &field1, &field2, &obj_type)!=3)
+            {
                 break;
             }
             //fprintf(fd, "%s\n", entry);
             elm = &(this->table[object_id]);
-            if (elm->type==0){ // skip if already set by next xreftable
+            if (elm->type==0)  // skip if already set by next xreftable
+            {
                 elm->major = object_id;
                 elm->type = obj_type=='f'? 0 : 1;
                 elm->offset = field1;
@@ -927,9 +1047,11 @@ bool ObjectTable:: get (MYFILE *f, size_t xref_poz, char *line, PdfObject *p_tra
             ++object_id;
             --object_count;
         }
-        else {
+        else
+        {
             int object_begin_tmp, object_count_tmp;
-            if (sscanf(entry,"%d %d", &object_begin_tmp, &object_count_tmp)!=2){
+            if (sscanf(entry,"%d %d", &object_begin_tmp, &object_count_tmp)!=2)
+            {
                 break;
             }
             object_id = object_begin_tmp;
@@ -938,15 +1060,19 @@ bool ObjectTable:: get (MYFILE *f, size_t xref_poz, char *line, PdfObject *p_tra
         }
     }
     //fclose(fd);
-    if (object_count!=0){
+    if (object_count!=0)
+    {
         return false;
     }
-	if (table[0].type!=0){//in some bad xref tables
+    if (table[0].type!=0) //in some bad xref tables
+    {
         debug("obj no 0 is not free");
-	}
-    while (!starts(line,"trailer")){
+    }
+    while (!starts(line,"trailer"))
+    {
         pos = myftell(f);
-        if (myfgets(line,LLEN,f,NULL)==EOF){
+        if (myfgets(line,LLEN,f,NULL)==EOF)
+        {
             message(ERROR, "trailer keyword not found");
             return false;
         }
@@ -955,8 +1081,9 @@ bool ObjectTable:: get (MYFILE *f, size_t xref_poz, char *line, PdfObject *p_tra
     // set seek pos just after trailer keyword
     myfseek(f, pos+8, SEEK_SET);
     if (p_trailer==NULL
-       || !p_trailer->read(f,this,NULL)
-       || p_trailer->type!=PDF_OBJ_DICT){
+            || !p_trailer->read(f,this,NULL)
+            || p_trailer->type!=PDF_OBJ_DICT)
+    {
         message(ERROR,"Could not read trailer dictionary");
         return false;
     }
@@ -969,10 +1096,10 @@ int ObjectTable:: addObject (PdfObject *obj)
     ObjectTableItem item = {NULL,0,0,0,0,0,0};
     table.resize(major+1, item);
 
-	table[major].major = major;
-	table[major].type = 1;
-	table[major].obj = obj;
-	return major;
+    table[major].major = major;
+    table[major].type = 1;
+    table[major].obj = obj;
+    return major;
 }
 
 PdfObject* ObjectTable:: getObject(int major, int minor)
@@ -985,56 +1112,66 @@ PdfObject* ObjectTable:: getObject(int major, int minor)
 
 void ObjectTable:: writeObjects (FILE *f)
 {
-	for (size_t i=1; i<table.size(); ++i){
-		switch (table[i].type){
-			case 0:
-				continue;
-			case 1:
-				table[i].offset = ftell(f);
-				if (fprintf(f,"%d %d obj\n", table[i].major, table[i].minor)<0){
-					message(FATAL,"I/O error");
-				}
-				if (table[i].obj->write(f)<0){
-					message(FATAL,"I/O error");
-				}
-				if (fprintf(f,"\nendobj\n")<0){
-					message(FATAL,"I/O error");
-				}
-				break;
-			default:
-				assert(0);
-		}
-	}
+    for (size_t i=1; i<table.size(); ++i)
+    {
+        switch (table[i].type)
+        {
+        case 0:
+            continue;
+        case 1:
+            table[i].offset = ftell(f);
+            if (fprintf(f,"%d %d obj\n", table[i].major, table[i].minor)<0)
+            {
+                message(FATAL,"I/O error");
+            }
+            if (table[i].obj->write(f)<0)
+            {
+                message(FATAL,"I/O error");
+            }
+            if (fprintf(f,"\nendobj\n")<0)
+            {
+                message(FATAL,"I/O error");
+            }
+            break;
+        default:
+            assert(0);
+        }
+    }
 }
 
 void ObjectTable:: writeXref (FILE *f)
 {
     table[0].type = 0;// obj 0 is always free
     table[0].minor = 65535; // and it has maximum gen id
-	fprintf(f, "xref\n%d %d\n", 0, table.size());
-	for (size_t i=0; i<table.size(); ++i){
+    fprintf(f, "xref\n%d %d\n", 0, table.size());
+    for (size_t i=0; i<table.size(); ++i)
+    {
         char type = (table[i].type) ? 'n' : 'f';
-		if (fprintf(f,"%010d %05d %c \n", table[i].offset, table[i].minor, type)<0){
-			message(FATAL, "I/O error");
-		}
-	}
+        if (fprintf(f,"%010d %05d %c \n", table[i].offset, table[i].minor, type)<0)
+        {
+            message(FATAL, "I/O error");
+        }
+    }
 }
 
 
 // *********** ------------- Token Parser ----------------- ***********
 
-typedef struct {
+typedef struct
+{
     char *str;
     size_t size;// allocated size
     size_t cpoz;//current position, str length+1
 } mystring;
 
-static int mystring_new(mystring * s){
+static int mystring_new(mystring * s)
+{
     s->size = 10;
     s->str = (char *) malloc(sizeof(char) * s->size);
     s->str[0] = 0;
     s->cpoz = 1;
-    if (s->str==NULL){
+    if (s->str==NULL)
+    {
         s->size = 0;
         return -1;
     }
@@ -1044,10 +1181,12 @@ static int mystring_new(mystring * s){
 static int mystring_add_char(mystring * s,char c)
 {
     char *new_str;
-    if (s->size == s->cpoz){
+    if (s->size == s->cpoz)
+    {
         s->size = s->size * 2;
         new_str = (char *) realloc(s->str, sizeof(char) * s->size);
-        if (new_str==NULL){
+        if (new_str==NULL)
+        {
             s->size = s->size / 2;
             return -1;
         }
@@ -1060,7 +1199,8 @@ static int mystring_add_char(mystring * s,char c)
 }
 
 
-Token:: Token() {
+Token:: Token()
+{
     type = TOK_UNKNOWN;
 }
 
@@ -1074,272 +1214,305 @@ Token:: get (MYFILE * f)
     double real_number, frac;
     // skip whitespace characters
     int newline = 0;
-    while (1){
+    while (1)
+    {
         c = mygetc(f);
-        switch (c){
-            case EOF:
-                this->type = TOK_EOF;
-                return true;
+        switch (c)
+        {
+        case EOF:
+            this->type = TOK_EOF;
+            return true;
             // white space
-            case CHAR_FF:
-            case CHAR_SP:
-            case CHAR_TAB:
-                newline = 0;
-                break;
+        case CHAR_FF:
+        case CHAR_SP:
+        case CHAR_TAB:
+            newline = 0;
+            break;
             // new line
-            case CHAR_LF:
-            case CHAR_CR:
-                newline = 1;
-                break;
-            default:
-                goto end_wh_sp;
+        case CHAR_LF:
+        case CHAR_CR:
+            newline = 1;
+            break;
+        default:
+            goto end_wh_sp;
         }
     }
 end_wh_sp:
     this->new_line = newline;
-    switch (c){
-        case '0':
-        case '1':
-        case '2':
-        case '3':
-        case '4':
-        case '5':
-        case '6':
-        case '7':
-        case '8':
-        case '9':
-        case '-':
-        case '+':
-        case '.':
-            if (c!='+' && c!='-'){// digit
-                number = c-'0';
-                this->sign = 0;
-            }
-            else {
-                number = 0;
-                switch (c) {
-                    case '+':
-                        this->sign = 1;
-                        break;
-                    case '-':
-                        minus = -1;
-                        this->sign = -1;
-                        break;
-                }
-            }
-            if (c=='.'){ // number beginning with '.' eg - .21
-                goto real_num;
-            }
-            while ((c=mygetc(f))!=EOF && isdigit(c)){
-                number = number*10+(c-'0');
-            }
-            switch (c){
-            // white spaces
-            case CHAR_FF:
-            case CHAR_SP:
-            case CHAR_TAB:
-            case CHAR_LF:
-            case CHAR_CR:
-            default:
-                myungetc(f);
-            case EOF:
-                this->type = TOK_INT;
-                this->integer = number * ((minus==-1)?-1:1);
-                return true;
-            case '.':
+    switch (c)
+    {
+    case '0':
+    case '1':
+    case '2':
+    case '3':
+    case '4':
+    case '5':
+    case '6':
+    case '7':
+    case '8':
+    case '9':
+    case '-':
+    case '+':
+    case '.':
+        if (c!='+' && c!='-') // digit
+        {
+            number = c-'0';
+            this->sign = 0;
+        }
+        else
+        {
+            number = 0;
+            switch (c)
+            {
+            case '+':
+                this->sign = 1;
+                break;
+            case '-':
+                minus = -1;
+                this->sign = -1;
                 break;
             }
-        real_num:
-            real_number = number;
-            frac = 10;
-            while ((c=mygetc(f))!=EOF && isdigit(c)){
-                real_number = real_number + (c-'0')/frac;
-                frac = frac * 10;
-            }
-            switch(c){
+        }
+        if (c=='.')  // number beginning with '.' eg - .21
+        {
+            goto real_num;
+        }
+        while ((c=mygetc(f))!=EOF && isdigit(c))
+        {
+            number = number*10+(c-'0');
+        }
+        switch (c)
+        {
+            // white spaces
+        case CHAR_FF:
+        case CHAR_SP:
+        case CHAR_TAB:
+        case CHAR_LF:
+        case CHAR_CR:
+        default:
+            myungetc(f);
+        case EOF:
+            this->type = TOK_INT;
+            this->integer = number * ((minus==-1)?-1:1);
+            return true;
+        case '.':
+            break;
+        }
+real_num:
+        real_number = number;
+        frac = 10;
+        while ((c=mygetc(f))!=EOF && isdigit(c))
+        {
+            real_number = real_number + (c-'0')/frac;
+            frac = frac * 10;
+        }
+        switch(c)
+        {
             /*bily znak*/
-            case CHAR_FF:
-            case CHAR_SP:
-            case CHAR_TAB:
+        case CHAR_FF:
+        case CHAR_SP:
+        case CHAR_TAB:
             // new line
-            case CHAR_LF:
-            case CHAR_CR:
-            case ']':
-            case '>':
-            case '/':
-                myungetc(f);
-            case EOF:
-                this->type = TOK_REAL;
-                this->real = real_number * ((minus==-1)?-1:1);
-                return true;
-            default:
-                this->type = TOK_UNKNOWN;
-                return false;
-            }
-            break;
-
-        case '[': /*begin array*/
-            this->type = TOK_BARRAY;
+        case CHAR_LF:
+        case CHAR_CR:
+        case ']':
+        case '>':
+        case '/':
+            myungetc(f);
+        case EOF:
+            this->type = TOK_REAL;
+            this->real = real_number * ((minus==-1)?-1:1);
             return true;
-        case ']': /*end array*/
-            this->type = TOK_EARRAY;
+        default:
+            this->type = TOK_UNKNOWN;
+            return false;
+        }
+        break;
+
+    case '[': /*begin array*/
+        this->type = TOK_BARRAY;
+        return true;
+    case ']': /*end array*/
+        this->type = TOK_EARRAY;
+        return true;
+    case '<': /*hexadecimal string or dictionary*/
+        if ((c=mygetc(f))=='<')
+        {
+            this->type = TOK_BDICT;
             return true;
-        case '<': /*hexadecimal string or dictionary*/
-            if ((c=mygetc(f))=='<'){
-                this->type = TOK_BDICT;
-                return true;
+        }
+        else
+        {
+            //this->str.type=PDF_STR_HEX;
+            mystring_new(&mstr);
+            mystring_add_char(&mstr, '<');
+            /*hexadecimal string*/
+            while (c!=EOF && c!='>')
+            {
+                mystring_add_char(&mstr,c);
+                c = mygetc(f);
             }
-            else {
-                //this->str.type=PDF_STR_HEX;
-                mystring_new(&mstr);
-                mystring_add_char(&mstr, '<');
-                /*hexadecimal string*/
-                while (c!=EOF && c!='>'){
-                    mystring_add_char(&mstr,c);
-                    c = mygetc(f);
-                }
-                if (c=='>') {
-                    mystring_add_char(&mstr, '>');
-                    mstr.str = (char*) realloc(mstr.str, mstr.cpoz);
-                    this->type = TOK_STR;
-                    this->str.len = mstr.cpoz-1;
-                    this->str.data = mstr.str;
-                    return true;
-                }
-                else {//EOF, should free mystring data
-                    this->type = TOK_UNKNOWN;
-                    return false;
-                }
-            }
-            break;
-
-        case '>': //end dictionary
-                if (mygetc(f)=='>'){
-                    this->type = TOK_EDICT;
-                    return true;
-                }
-                else {
-                    this->type = TOK_UNKNOWN;
-                    myungetc(f);
-                    return false;
-                }
-            break;
-        case '(': // literal string, it may contain balanced parentheses
-                parenthes = 0;
-                //this->str.type=PDF_STR_CHR;
-                mystring_new(&mstr);
-                mystring_add_char(&mstr, '(');
-                while ((c=mygetc(f))!=EOF){
-                    switch(c){
-                        case '\\':
-                            mystring_add_char(&mstr,c);
-                            c=mygetc(f);
-
-                            break;
-                        case '(':
-                            parenthes++;
-                            break;
-                        case ')':
-                            if (parenthes==0){
-                                goto end_lit_str;
-                            }
-                            --parenthes;
-                            break;
-                    }
-                    mystring_add_char(&mstr,c);
-                }
-end_lit_str:
-                if (c==EOF){//should free mystring data here
-                    this->type = TOK_UNKNOWN;
-                    return false;
-                }
-                mystring_add_char(&mstr, ')');
+            if (c=='>')
+            {
+                mystring_add_char(&mstr, '>');
                 mstr.str = (char*) realloc(mstr.str, mstr.cpoz);
                 this->type = TOK_STR;
                 this->str.len = mstr.cpoz-1;
                 this->str.data = mstr.str;
                 return true;
-            break;
-        case '/':  //name object
-            i=0;
-            while ((c=mygetc(f))!=EOF){
-                switch(c){
-                    case CHAR_FF:
-                    case CHAR_SP:
-                    case CHAR_TAB:
-                    case CHAR_LF:
-                    case CHAR_CR:
-                    case '<':
-                    case '>':
-                    case '{':
-                    case '}':
-                    case '/':
-                    case '%':
-                    case '(':
-                    case ')':
-                    case '[':
-                    case ']':
-                        myungetc(f);
-                        goto end_name;
-                }
-                if (i+1<PDF_NAME_MAX_LEN){
-                    this->name[i] = c;
-                    ++i;
-                }
-                else {
-                    break;
-                }
             }
-end_name:
-            this->name[i] = 0;
-            this->type = TOK_NAME;
-            return true;
-        case '%': //comment, skip characters to end of line, then find next token
-            while ((c=mygetc(f))!=EOF && c!=CHAR_LF && c!=CHAR_CR)
-                ;
-            if (c==EOF){
+            else  //EOF, should free mystring data
+            {
                 this->type = TOK_UNKNOWN;
-                return true;
+                return false;
             }
-            else {
-                myungetc(f);
-                return this->get(f);
-            }
-            break;
-        default:
-            i=0;
-            do {
-            switch (c){
-                case CHAR_FF:
-                case CHAR_SP:
-                case CHAR_TAB:
-                case CHAR_LF:
-                case CHAR_CR:
-                case '<':
-                case '>':
-                case '{':
-                case '}':
-                case '/':
-                case '%':
-                case '(':
-                case ')':
-                case '[':
-                case ']':
-                    myungetc(f);
-                    goto end_id;
-                }
-                if (i+1<PDF_ID_MAX_LEN){
-                    this->id[i] = c;
-                    ++i;
-                }
-                else {
-                    break;
-                }
-            } while ((c=mygetc(f))!=EOF);
-end_id:
-            this->id[i] = 0;
-            this->type = TOK_ID;
+        }
+        break;
+
+    case '>': //end dictionary
+        if (mygetc(f)=='>')
+        {
+            this->type = TOK_EDICT;
             return true;
+        }
+        else
+        {
+            this->type = TOK_UNKNOWN;
+            myungetc(f);
+            return false;
+        }
+        break;
+    case '(': // literal string, it may contain balanced parentheses
+        parenthes = 0;
+        //this->str.type=PDF_STR_CHR;
+        mystring_new(&mstr);
+        mystring_add_char(&mstr, '(');
+        while ((c=mygetc(f))!=EOF)
+        {
+            switch(c)
+            {
+            case '\\':
+                mystring_add_char(&mstr,c);
+                c=mygetc(f);
+
+                break;
+            case '(':
+                parenthes++;
+                break;
+            case ')':
+                if (parenthes==0)
+                {
+                    goto end_lit_str;
+                }
+                --parenthes;
+                break;
+            }
+            mystring_add_char(&mstr,c);
+        }
+end_lit_str:
+        if (c==EOF) //should free mystring data here
+        {
+            this->type = TOK_UNKNOWN;
+            return false;
+        }
+        mystring_add_char(&mstr, ')');
+        mstr.str = (char*) realloc(mstr.str, mstr.cpoz);
+        this->type = TOK_STR;
+        this->str.len = mstr.cpoz-1;
+        this->str.data = mstr.str;
+        return true;
+        break;
+    case '/':  //name object
+        i=0;
+        while ((c=mygetc(f))!=EOF)
+        {
+            switch(c)
+            {
+            case CHAR_FF:
+            case CHAR_SP:
+            case CHAR_TAB:
+            case CHAR_LF:
+            case CHAR_CR:
+            case '<':
+            case '>':
+            case '{':
+            case '}':
+            case '/':
+            case '%':
+            case '(':
+            case ')':
+            case '[':
+            case ']':
+                myungetc(f);
+                goto end_name;
+            }
+            if (i+1<PDF_NAME_MAX_LEN)
+            {
+                this->name[i] = c;
+                ++i;
+            }
+            else
+            {
+                break;
+            }
+        }
+end_name:
+        this->name[i] = 0;
+        this->type = TOK_NAME;
+        return true;
+    case '%': //comment, skip characters to end of line, then find next token
+        while ((c=mygetc(f))!=EOF && c!=CHAR_LF && c!=CHAR_CR)
+            ;
+        if (c==EOF)
+        {
+            this->type = TOK_UNKNOWN;
+            return true;
+        }
+        else
+        {
+            myungetc(f);
+            return this->get(f);
+        }
+        break;
+    default:
+        i=0;
+        do
+        {
+            switch (c)
+            {
+            case CHAR_FF:
+            case CHAR_SP:
+            case CHAR_TAB:
+            case CHAR_LF:
+            case CHAR_CR:
+            case '<':
+            case '>':
+            case '{':
+            case '}':
+            case '/':
+            case '%':
+            case '(':
+            case ')':
+            case '[':
+            case ']':
+                myungetc(f);
+                goto end_id;
+            }
+            if (i+1<PDF_ID_MAX_LEN)
+            {
+                this->id[i] = c;
+                ++i;
+            }
+            else
+            {
+                break;
+            }
+        }
+        while ((c=mygetc(f))!=EOF);
+end_id:
+        this->id[i] = 0;
+        this->type = TOK_ID;
+        return true;
     }
     return true;
 }
@@ -1347,12 +1520,13 @@ end_id:
 void
 Token:: freeData()
 {
-    switch (this->type) {
-        case TOK_STR:
-            free(this->str.data);
-            break;
-        default:
-            break;
+    switch (this->type)
+    {
+    case TOK_STR:
+        free(this->str.data);
+        break;
+    default:
+        break;
     }
 }
